+++ lib/src/data/provider/auth_service_provider.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_riverpod/legacy.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import 'package:monami/src/presentation/views/view_models/base_model.dart';
 
-final authServiceProvider = ChangeNotifierProvider((ref) => BaseViewModel());
+// final authServiceProvider = ChangeNotifierProvider((ref) => BaseViewModel());
