+++ lib/src/data/provider/user_id_provider.dart
@@ -1,7 +1,7 @@
-import 'package:flutter_riverpod/legacy.dart' show ChangeNotifierProvider;
+// import 'package:flutter_riverpod/legacy.dart' show ChangeNotifierProvider;
 
-import 'package:monami/src/presentation/views/view_models/base_model.dart';
+// import 'package:monami/src/presentation/views/view_models/base_model.dart';
 
-final userId = ChangeNotifierProvider((ref) => UserIdViewModel());
+// final userId = ChangeNotifierProvider((ref) => UserIdViewModel());
 
-class UserIdViewModel extends BaseViewModel {}
+// class UserIdViewModel extends BaseViewModel {}
