+++ lib/src/features/auth/providers/auth_state_provider.dart
@@ -1,6 +1,6 @@
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:monami/src/presentation/views/view_models/base_model.dart';
+// import 'package:hooks_riverpod/hooks_riverpod.dart';
+// import 'package:monami/src/presentation/views/view_models/base_model.dart';
 
-final authenticationProvider = Provider<BaseViewModel>((ref) {
-  return BaseViewModel();
-});
+// final authenticationProvider = Provider<BaseViewModel>((ref) {
+//   return BaseViewModel();
+// });
