+++ lib/src/presentation/views/login/login_state.dart
@@ -1,0 +1,19 @@
+class LoginState {
+  final bool isLoading;
+  final String? error;
+
+  const LoginState({
+    this.isLoading = false,
+    this.error,
+  });
+
+  LoginState copyWith({
+    bool? isLoading,
+    String? error,
+  }) {
+    return LoginState(
+      isLoading: isLoading ?? this.isLoading,
+      error: error,
+    );
+  }
+}
