+++ lib/src/presentation/views/profile/components/loading_view.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class LoadingView extends StatelessWidget {
+  const LoadingView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      backgroundColor: Color(0xFFF8F9FA),
+      body: Center(
+        child: CircularProgressIndicator(
+          valueColor: AlwaysStoppedAnimation(Color(0xFF667EEA)),
+        ),
+      ),
+    );
+  }
+}
+
+class ErrorView extends StatelessWidget {
+  const ErrorView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      backgroundColor: Color(0xFFF8F9FA),
+      body: Center(
+        child: Text('Could not load profile'),
+      ),
+    );
+  }
+}
