+++ lib/src/presentation/views/splash/splash_view_model.dart
@@ -1,16 +1,21 @@
-import 'package:flutter_riverpod/legacy.dart' show ChangeNotifierProvider;
+import 'dart:developer';
+
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:monami/src/presentation/views/view_models/base_model.dart';
 import 'package:monami/src/utils/router/route_name.dart';
 
-final splashViewModelProvider = ChangeNotifierProvider((ref) {
-  return SplashViewModel();
-});
+final splashViewModelProvider =
+    NotifierProvider<SplashViewModel, bool>(SplashViewModel.new);
 
-class SplashViewModel extends BaseViewModel {
-  SplashViewModel() {
+class SplashViewModel extends Notifier<bool> with BaseViewModel {
+  @override
+  bool build() {
     checkLoginStatus();
+    return false;
   }
+
   Future<void> checkLoginStatus() async {
+    state = true;
     try {
       await Future.delayed(const Duration(milliseconds: 1500));
 
@@ -26,10 +31,8 @@ class SplashViewModel extends BaseViewModel {
         navigationHandler.pushReplacementNamed(Routes.onboardingViewRoute);
       }
     } catch (e) {
-      log(e);
-      await Future.delayed(const Duration(milliseconds: 800));
-      // Skip login on error too - go directly to home for UI testing
-      navigationHandler.pushReplacementNamed(Routes.homeViewRoute);
+      state = false;
+      log(e.toString());
     }
   }
 }
